cleanup vi code correct a mistake

+++ bld/vi/h/rtns.h
@@ -765,7 +765,7 @@ vi_rc   ExpandWhiteSpace( void );
 vi_rc   CompressWhiteSpace( void );
 bool    ExpandTabsInABufferUpToColumn( int, char *, int, char *, int );
 bool    ExpandTabsInABuffer( const char *, int, char *, int );
-int     InsertTabSpace( int, char *, bool * );
+size_t  InsertTabSpace( size_t, char *, bool * );
 int     GetVirtualCursorPosition( char *, int );
 int     VirtualColumnOnCurrentLine( int );
 int     RealColumnOnCurrentLine( int );
